refactor(transform): index callable provenance
